Fix warning: implicit conversion changes signedness
visp/tutorial/segmentation/color/tutorial-hsv-range-tuner.cpp:183:15: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  183 |     width = I.getWidth();
      |           ~ ~~^~~~~~~~~~
visp/tutorial/segmentation/color/tutorial-hsv-range-tuner.cpp:184:16: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  184 |     height = I.getHeight();
      |            ~ ~~^~~~~~~~~~~
visp/tutorial/segmentation/color/tutorial-hsv-range-tuner.cpp:208:39: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  208 |   vpImage<unsigned char> mask(height, width);
      |                          ~~~~         ^~~~~
visp/tutorial/segmentation/color/tutorial-hsv-range-tuner.cpp:208:31: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  208 |   vpImage<unsigned char> mask(height, width);
      |                          ~~~~ ^~~~~~
visp/tutorial/segmentation/color/tutorial-hsv-range-tuner.cpp:209:39: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  209 |   vpImage<vpRGBa> I_segmented(height, width);
      |                   ~~~~~~~~~~~         ^~~~~
visp/tutorial/segmentation/color/tutorial-hsv-range-tuner.cpp:209:31: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  209 |   vpImage<vpRGBa> I_segmented(height, width);
      |                   ~~~~~~~~~~~ ^~~~~~
visp/tutorial/segmentation/color/tutorial-hsv-range-tuner.cpp:214:103: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  214 |   std::shared_ptr<vpDisplay> d_I_segmented = vpDisplayFactory::createDisplay(I_segmented, I.getWidth()+75, 0, "Segmented frame");
      |                                              ~~~~~~~~~~~~~~~~                             ~~~~~~~~~~~~^~~

+++ tutorial/segmentation/color/tutorial-hsv-range-tuner.cpp
@@ -153,7 +153,7 @@ int main(int argc, const char *argv[])
   hsv_values_trackbar[5] = max_value;   // High V
 
   vpImage<vpRGBa> I;
-  int width, height;
+  unsigned int width, height;
 
 #if defined(VISP_HAVE_REALSENSE2)
   vpRealSense2 rs;
@@ -164,7 +164,7 @@ int main(int argc, const char *argv[])
     width = 848; height = 480;
     int fps = 60;
     rs2::config config;
-    config.enable_stream(RS2_STREAM_COLOR, width, height, RS2_FORMAT_RGBA8, fps);
+    config.enable_stream(RS2_STREAM_COLOR, static_cast<int>(width), static_cast<int>(height), RS2_FORMAT_RGBA8, fps);
     config.disable_stream(RS2_STREAM_DEPTH);
     config.disable_stream(RS2_STREAM_INFRARED, 1);
     config.disable_stream(RS2_STREAM_INFRARED, 2);
@@ -211,7 +211,7 @@ int main(int argc, const char *argv[])
 #if (VISP_CXX_STANDARD >= VISP_CXX_STANDARD_11)
   vpImage<vpHSV<unsigned char, true>> Ihsv;
   std::shared_ptr<vpDisplay> d_I = vpDisplayFactory::createDisplay(I, 0, 0, "Current frame");
-  std::shared_ptr<vpDisplay> d_I_segmented = vpDisplayFactory::createDisplay(I_segmented, I.getWidth()+75, 0, "Segmented frame");
+  std::shared_ptr<vpDisplay> d_I_segmented = vpDisplayFactory::createDisplay(I_segmented, static_cast<int>(I.getWidth())+75, 0, "Segmented frame");
 #else
   vpImage<unsigned char> H(height, width);
   vpImage<unsigned char> S(height, width);
